Rename lti app to bridge_lti due to pylti lib shadowing
