+++ bridge-adaptivity/adaptive_edx/urls.py
@@ -7,7 +7,7 @@ urlpatterns = [
     url(r'^$', views.home, name='home'),
     url(r'^auth_error/', views.lti_auth_error, name='lti_auth_error'),
     url(r'^admin/', include(admin.site.urls)),
-    url(r'^lti/', include('lti.urls', namespace="lti")),
+    url(r'^lti/', include('bridge_lti.urls', namespace="bridge-lti")),
     url(r'^module/', include('module.urls', namespace="module")),
     url(r'^api/', include('api.urls', namespace="api")),
     # url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework')),
